requirements.txt
fastapi
uvicorn
pydantic
python-multipart
aiofiles
psycopg2-binary
